final-pjt-back/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==1.1.0
idna==3.4
inflection==0.5.1
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jsonschema==4.17.0
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyrsistent==0.19.2
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
threadpoolctl==3.1.0
traitlets==5.4.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
